yet another buggy version check

+++ install_vapoursynth_libs.py
@@ -106,7 +106,7 @@ else:
 		runCmd("cp {0} ../bin".format("VapourSynth.dll"))
 		
 		pydName = "vapoursynth.cp36-win_amd64.pyd"
-		if int(ver_suff) > 44: # Version 45 and above use python 3.7		
+		if float(ver_suff) > 44: # Version 45 and above use python 3.7		
 			pydName = "vapoursynth.cp37-win_amd64.pyd"
 			VSS_PC = VSS_PC.replace("%%PY_VER_DOT%%","3.7").replace("%%PY_VER%%","37")
 		else:
